backend/requirements.txt
fastapi>=0.135,<1.0
uvicorn[standard]>=0.42,<1.0
pydantic-settings>=2.13,<3.0
cachetools>=7.0,<8.0
pandas>=2.3,<3.0
openai>=2.30,<3.0
databricks-sdk>=0.102,<1.0
databricks-connect>=15.4,<17.0
sentry-sdk[fastapi]>=2.0,<3.0
pytest>=9.0,<10.0
